third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-preventsilentaccess-basics.html

<!DOCTYPE html>
<title>Credential Manager: preventSilentAccess() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/webauth/authenticator.mojom.js"></script>
<script src="resources/authenticator-helpers.js"></script>
<script>
function stubResolverChecker(c) {
    assert_equals(c, undefined);
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("preventSilentAccess() should not reject, but did: " + reason.name);
}

async_test(function () {
    navigator.credentials.preventSilentAccess().then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of preventSilentAccess()");
</script>
